refactor: move `LabelledPrecendenceGraph` from `Ast` to `shared`
